UX - First draft of multiple template support in preparation for supporting 300 and 600ppi templates

+++ src/main/java/com/mickeytheq/hades/strangeeons/tasks/GenerateComparisonImagesTaskAction.java
@@ -107,7 +107,7 @@ public class GenerateComparisonImagesTaskAction extends BaseTaskAction {
         }
     }
 
-    private List<BufferedImage> createHadesImages(Path cardPath, ProjectContext projectContext, RenderTarget renderTarget, double dpi) {
+    private List<BufferedImage> createHadesImages(Path cardPath, ProjectContext projectContext, RenderTarget renderTarget, int dpi) {
         if (!Files.exists(cardPath))
             return Lists.newArrayList(null, null);
